+++ src/components/Drawer.jsx
@@ -16,8 +16,6 @@ import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
 import InboxIcon from '@material-ui/icons/MoveToInbox';
 import MailIcon from '@material-ui/icons/Mail';
-import SampleCard from './Card';
-import FullWidthGrid from './Grid';
 
 const drawerWidth = 240;
 
@@ -189,11 +187,6 @@ export default function MiniDrawer() {
             </Drawer>
             <main className={classes.content}>
                 <div className={classes.toolbar} />
-                <FullWidthGrid>
-                    <SampleCard />
-                    <SampleCard />
-                    <SampleCard />
-                </FullWidthGrid>
             </main>
         </div>
     );
